Two missed flight recorder configs #21205

+++ akka-multi-node-testkit/src/main/scala/akka/remote/testkit/FlightRecordingSupport.scala
@@ -35,7 +35,7 @@ trait FlightRecordingSupport { self: MultiNodeSpec ⇒
    * Delete flight the recorder file if it exists
    */
   final protected def deleteFlightRecorderFile(): Unit = {
-    if (arteryEnabled && Files.exists(flightRecorderFile)) {
+    if (arteryEnabled && destinationIsValidForDump() && Files.exists(flightRecorderFile)) {
       Files.delete(flightRecorderFile)
     }
   }
